+++ src/ChessDotCpp/constants.h
@@ -19,7 +19,7 @@ public:
 
     static constexpr Score Inf = 32000;
     static constexpr Score Mate = 31000;
-    static constexpr Score TablebaseMate = 30500;
+    static constexpr Score TablebaseMate = 20000;
     static constexpr Score MateThreshold = 30000;
     static constexpr Score EndgameMaterial = 1300;